% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicdworkshop-package.R
\docType{package}
\name{cicdworkshop-package}
\alias{cicdworkshop}
\alias{cicdworkshop-package}
\title{cicdworkshop: Demo Package for R/Pharma CI/CD Workshop}
\description{
Intended to demo CI/CD pipelines on GitHub.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/boxizhang/CICD}
}

}
\author{
\strong{Maintainer}: Open Source \email{osauthor@olink.com}

}
\concept{internal}
\keyword{internal}
